fix: address PR review feedback on concurrency doc and test portability
Move TestIndexerConcurrentPartialFailure to indexer_unix_test.go with
//go:build !windows tag since it uses os.Getuid() and POSIX file
permissions. Tighten Indexer doc comment to clarify that only store
writes are serialized and injected Chunker/VectorStore must be
concurrency-safe.
